Split test file
The test file has been split into smaller test files.
